@@ -4,17 +4,17 @@ plugins {
     id "org.jetbrains.kotlin.jvm" version "1.7.10"
     id "org.jetbrains.kotlin.plugin.jpa" version "1.7.10"
     id "org.jetbrains.kotlin.plugin.spring" version "1.7.10"
-    id "org.springframework.boot" version "2.7.5"
-    id "io.spring.dependency-management" version "1.0.11.RELEASE"
+    id "org.springframework.boot" version "3.0.6"
+    id "io.spring.dependency-management" version "1.1.0"
     id "com.gorylenko.gradle-git-properties" version "2.4.1"
     id "com.adarshr.test-logger" version "3.2.0"
 }
 
-[compileKotlin, compileTestKotlin]*.kotlinOptions*.jvmTarget = "11"
+[compileKotlin, compileTestKotlin]*.kotlinOptions*.jvmTarget = "17"
 
 ext {
     jwtVersion = "4.2.1"
-    openApiVersion = "1.6.12"
+    openApiVersion = "2.1.0"
     selenideVersion = "6.7.2"
     selenoidAdapterVersion = "2.3.3"
     postgresContainerVersion = "1.17.0"
@@ -60,8 +60,8 @@ dependencies {
     implementation "org.flywaydb:flyway-core"
 
     implementation "io.micrometer:micrometer-registry-prometheus"
-    implementation "org.springdoc:springdoc-openapi-ui:$openApiVersion"
-    implementation "org.springdoc:springdoc-openapi-kotlin:$openApiVersion"
+    implementation "org.springdoc:springdoc-openapi-starter-common:$openApiVersion"
+    implementation "org.springdoc:springdoc-openapi-starter-webmvc-ui:$openApiVersion"
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
     implementation "org.jetbrains.kotlin:kotlin-reflect"
     implementation "com.fasterxml.jackson.module:jackson-module-kotlin"
