@@ -13,7 +13,7 @@ typealias CustomPostgresContainer = PostgreSQLContainer<*>
 @TestConfiguration
 class DatabaseTestConfiguration {
     companion object {
-        private const val POSTGRES_IMAGE = "postgres:13-alpine"
+        private const val POSTGRES_IMAGE = "postgres:15-alpine"
         private const val DATABASE_NAME = "todo_list"
         private const val USERNAME = "program"
         private const val PASSWORD = "test"
